move inmem secrets to crate and impl in lambda
