--- runtimes/aws-lambda/host/.cargo/config.toml
@@ -1,4 +1,0 @@
-[target.'cfg(target_os = "linux")']
-rustflags = [
-  "-C", "link-arg=-Wl,-E"
-]
