+++ runtimes/components/wasi-secrets/in-memory/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "assemblylift-wasi-secrets-in-memory"
+version = "0.0.0"
+description = "A simple in-memory secrets manager to develop wasi-secrets"
+edition = "2021"
+
+[dependencies]
+anyhow = "1"
+chacha20poly1305 = "0.10"
+once_cell = "1.17"
+rand = "0.8"
+tracing = "0.1"
+
+assemblylift-core = { version = "0.4.0-beta.0", path = "../../../../core" }
